Improve twig extensions
