+++ src/Extension/BaseExtension.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Sonata\FormatterBundle\Extension;
 
-use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 
 abstract class BaseExtension extends AbstractExtension implements ExtensionInterface
@@ -41,10 +40,6 @@ abstract class BaseExtension extends AbstractExtension implements ExtensionInter
     public function getAllowedMethods(): array
     {
         return [];
-    }
-
-    public function initRuntime(Environment $environment): void
-    {
     }
 
     public function getTokenParsers(): array
