+++ tests/App/AppKernel.php
@@ -111,6 +111,6 @@ final class AppKernel extends Kernel
 
     private function getBaseDir(): string
     {
-        return sprintf('%s/sonata-formatter-bundle/var/', sys_get_temp_dir());
+        return sprintf('%s/sonata-fdsafdasformavcxfdsgggfdfdsttefdsfdsafsdr-bundle/var/', sys_get_temp_dir());
     }
 }
