+++ tests/App/Resources/views/text_formatter_test.html.twig
@@ -9,7 +9,7 @@ file that was distributed with this source code.
 
 #}
 
-{%- apply spaceless|format_text('rawhtml')|raw %}
+{%- apply spaceless|format_text('rawhtml') %}
     <h1>Test</h1>
     <% for i in 1..5 %>
         <% if i > 2 %>
